fix: NoopDataSource.open must return Long not Int
CI caught one more type error in NoopDataSource:

  e: CachedSongExporter.kt:165:55
    Return type mismatch: expected 'Long', actual 'Int'.

DataSource.open() returns Long (verified against MusicService.kt:7840
which overrides the same method with the same signature), but
C.LENGTH_UNSET in androidx.media3.common.C is declared as Int.
C.RESULT_END_OF_INPUT is also Int but is used as the return of read()
which returns Int, so that one is fine.

Wrap C.LENGTH_UNSET in .toLong() so the open() override type-checks.

+++ app/src/main/kotlin/dev/citali/lunartune/playback/CachedSongExporter.kt
@@ -162,7 +162,7 @@ class CachedSongExporter(
      * cache gaps no longer throw — they just produce a truncated export instead of a failure.
      */
     private object NoopDataSource : DataSource {
-        override fun open(dataSpec: DataSpec): Long = C.LENGTH_UNSET
+        override fun open(dataSpec: DataSpec): Long = C.LENGTH_UNSET.toLong()
         override fun read(
             buffer: ByteArray,
             offset: Int,
